re-organize moment matching into separate modules for better clarity
